Tue, Jan 14, 2025, 9:03 PM -08:00

+++ _python/trading.py
@@ -10,8 +10,8 @@ while True:
     elif entry < stoploss:
         target = entry - (stoploss - entry)
 
-    risk_2 = ((balance / 100) * 2) * target
-    risk_15 = ((balance / 100) * 1.5) * target
-    risk_1 = (balance / 100) * target
+    risk_2 = int(((balance / 100) * 2) * target)
+    risk_15 = int(((balance / 100) * 1.5) * target)
+    risk_1 = int((balance / 100) * target)
 
     print(f"{ticker} --> Target: {target}  Entry: {entry}  Stoploss: {stoploss}    Risk? 2%: -{risk_2}- 1.5%: -{risk_15}- 1%: -{risk_1}-")
